+++ src/ray_casting/vertical_wall_distance.c
@@ -6,19 +6,19 @@
 /*   By: alida-si <alida-si@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/29 11:39:01 by alida-si          #+#    #+#             */
-/*   Updated: 2023/04/03 00:11:04 by alida-si         ###   ########.fr       */
+/*   Updated: 2023/04/03 09:58:02 by alida-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-float	distance_to_vertical_wall_right(t_game *game, float angle)
+double	distance_to_vertical_wall_right(t_game *game, double angle)
 {
-	float	y_coordinate;
-	float	x_coordinate;
-	float	x_step;
-	float	y_step;
-	float	ray_length;
+	double	y_coordinate;
+	double	x_coordinate;
+	double	x_step;
+	double	y_step;
+	double	ray_length;
 
 	x_coordinate = ((int)game->player->x_position / MAP_CELL) * MAP_CELL + MAP_CELL;
 	y_coordinate = ((int)(game->player->x_position - x_coordinate))
@@ -29,13 +29,13 @@ float	distance_to_vertical_wall_right(t_game *game, float angle)
 	return (ray_length);
 }
 
-float	distance_to_vertical_wall_left(t_game *game, float angle)
+double	distance_to_vertical_wall_left(t_game *game, double angle)
 {
-	float	y_coordinate;
-	float	x_coordinate;
-	float	x_step;
-	float	y_step;
-	float	ray_length;
+	double	y_coordinate;
+	double	x_coordinate;
+	double	x_step;
+	double	y_step;
+	double	ray_length;
 
 	x_coordinate = ((int)game->player->x_position / MAP_CELL) * MAP_CELL - 1;
 	y_coordinate = ((int)(game->player->x_position - x_coordinate))
